Deletes RequestHandler
