+++ src/main/java/httpserver/server/HTTPServer.java
@@ -1,7 +1,7 @@
 package httpserver.server;
 
 import httpserver.ClientSocket;
-import httpserver.RequestHandler;
+import httpserver.HTTPRouter;
 import httpserver.SocketServer;
 import httpserver.httprequests.HTTPRequest;
 import httpserver.httpresponse.HTTPResponse;
@@ -9,17 +9,17 @@ import httpserver.httpresponse.HTTPResponse;
 public class HTTPServer {
 
     private final SocketServer socketServer;
-    private final RequestHandler requestHandler;
+    private final HTTPRouter router;
 
-    public HTTPServer(SocketServer socketServer, RequestHandler requestHandler) {
+    public HTTPServer(SocketServer socketServer, HTTPRouter router) {
         this.socketServer = socketServer;
-        this.requestHandler = requestHandler;
+        this.router = router;
     }
 
     public void start() {
         ClientSocket socket = socketServer.serve();
         HTTPRequest httpRequest = socket.getRequest();
-        HTTPResponse httpResponse = requestHandler.handle(httpRequest);
+        HTTPResponse httpResponse = router.route(httpRequest);
         socket.sendResponse(httpResponse);
         socket.close();
     }
