+++ src/test/java/httpserver/server/HTTPServerTest.java
@@ -1,9 +1,8 @@
 package httpserver.server;
 
-import httpserver.RequestHandler;
+import httpserver.HTTPRouter;
 import httpserver.httprequests.HTTPRequest;
 import httpserver.httpresponse.HTTPResponse;
-import httpserver.httpresponses.ResponseDateFake;
 import org.junit.Test;
 
 import java.io.IOException;
@@ -16,18 +15,18 @@ public class HTTPServerTest {
 
     @Test
     public void canAcceptConnectionsOnThePortWhenListening() throws IOException {
-        HTTPServer httpServer = new HTTPServer(connectionsSocketServer, new RequestHandlerFake());
+        HTTPServer httpServer = new HTTPServer(connectionsSocketServer, new RouterDummy());
 
         httpServer.start();
 
         assertEquals(1, connectionsSocketServer.getConnections());
     }
 
-    private class RequestHandlerFake implements RequestHandler {
+    private class RouterDummy implements HTTPRouter {
 
         @Override
-        public HTTPResponse handle(HTTPRequest httpRequest) {
-            return new HTTPResponse(200, "OK", new ResponseDateFake());
+        public HTTPResponse route(HTTPRequest httpRequest) {
+            return null;
         }
     }
 }
